not completed

+++ ft_putchar_fd_count.c
@@ -6,13 +6,13 @@
 /*   By: sakitaha <sakitaha@student.42tokyo.jp>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/05/21 00:53:02 by sakitaha          #+#    #+#             */
-/*   Updated: 2023/06/21 21:08:44 by sakitaha         ###   ########.fr       */
+/*   Updated: 2023/06/21 22:09:59 by sakitaha         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "ft_printf.h"
 
-int	ft_putchar_fd_count(char c, int fd)
+ssize_t	ft_putchar_fd_count(char c, int fd)
 {
 	return (write(fd, &c, 1));
 }
